+++ diff/diff.go
@@ -6,7 +6,6 @@ package diff
 
 import (
 	"io"
-	"io/ioutil"
 
 	"github.com/dsnet/compress/bzip2"
 	"github.com/icedream/go-bsdiff/internal"
@@ -14,11 +13,11 @@ import (
 )
 
 func Diff(oldReader, newReader io.Reader, patchWriter io.Writer) (err error) {
-	oldBytes, err := ioutil.ReadAll(oldReader)
+	oldBytes, err := io.ReadAll(oldReader)
 	if err != nil {
 		return
 	}
-	newBytes, err := ioutil.ReadAll(newReader)
+	newBytes, err := io.ReadAll(newReader)
 	if err != nil {
 		return
 	}